Move existing File IO into the FileIO module
